_code-samples/confidential-transfers/py/requirements.txt
xrpl-py>=4.0.0
-e /Users/mshodunke/Ripple/xrpl-apps/py-cmpt

